requirements.txt
Flask
paho-mqtt
APScheduler
python-dotenv
